requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cloudinary==1.42.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Markdown==3.7
PyJWT==2.10.1
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
